# backend/internal/docs/swagger.yaml
basePath: /api/
definitions:
  _ResponseWithData:
    description: Общий ответ success/error, содержащий произвольные данные.
    properties:
      data:
        description: Объект полезной нагрузки
      status:
        description: Результат запроса
        type: string
    type: object
  _ResponseWithMessage:
    description: Общий простой ответ, который передает только понятное для человека
      сообщение.
    properties:
      message:
        description: Человеко-читаемое сообщение
        type: string
      status:
        description: Результат запроса
        type: string
    type: object
...
  CheckRequestRequest:
    description: описание одной проверки
    properties:
      params:
        additionalProperties: true
        description: Params параметры проверки
        type: object
      type:
        description: 'Type тип проверки: http|ping|tcp|traceroute|dns'
        example: http
        type: string
    required:
    - params
    - type
    type: object
...
  TaskMessageRequest:
    description: Task для выполнения проверок сети (HTTP, ping, TCP, traceroute, DNS)
    properties:
      broadcast:
        description: Отправлять ли запрос на агенты всех регионов или берётся ближайший
          1 агент к клиенту
        example: false
        type: boolean
      checks:
        description: Checks список проверок
        items:
          $ref: '#/definitions/CheckRequestRequest'
        type: array
      target:
        description: Target домен или IP, который нужно проверить
        example: example.com
        type: string
      timeoutSeconds:
        description: TimeoutSeconds время выполнения всех задачи в секундах
        example: 20
        type: integer
    required:
    - checks
    - target
    - timeoutSeconds
    type: object
...
  hackathon-back_internal_model.CheckResultResponse:
    properties:
      agentId:
        type: string
      agentRegion:
        type: string
      finishedAt:
        type: string
      payload:
        items:
          type: integer
        type: array
      requestId:
        type: string
      startedAt:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
...
  hackathon-back_internal_model.Request:
    properties:
      broadcast:
        type: boolean
      checkTypes:
        items:
          type: string
        type: array
      clientASN:
        type: integer
      clientCC:
        type: string
      clientIP:
        type: string
      clientRegion:
        type: string
      createdAt:
        type: string
      id:
        type: string
      requestJSON:
        items:
          type: integer
        type: array
      status:
        type: string
      target:
        type: string
      timeoutSeconds:
        type: integer
      updatedAt:
        type: string
      userAgent:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    Флоу авторизации: сначала пользователь регистрируется, в ответе получает модельку своего user и токен.
    Токен + код с почты он отправляет на ручку confirm, если что-то идёт не так (токен истёк, не пришёл код на почту), то запрос нужно отправить на ручку resend-confirmation.
    Далее уже можно авторизоваться, login/refresh/test-login выставляет в cookie access и refresh токены, фронтенду, ничего с ними делать не нужно, они сами по себе живут в браузере и отправляются при каждом запросе.
    Специально для мобильного приложения при login/refresh/test-login токены дублируются в теле ответа.
    При запросе к защищённым ручкам API мобильному приложению необходимо выставить заголовок Authorization: Bearer *access_token*.
    При refresh мобильное приложение передаёт refresh токен в теле запроса.
  title: Hackathon API
  version: 0.1.0
paths:
...
  /check/{id}:
    get:
      description: Возвращает текущее состояние проверок по request_id одним ответом.
      parameters:
      - description: Request UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/_ResponseWithData'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/hackathon-back_internal_model.CheckResultResponse'
                  type: array
              type: object
        "400":
          description: Invalid path param
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "500":
          description: Failed to get results
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
      summary: Получить результаты сетевых проверок.
      tags:
      - Checks
  /check/task:
    post:
      consumes:
      - application/json
      description: Принимает проверки, достаёт IP клиента, определяет регион, пишет
        в checks=PENDING и в outbox кладёт.
      parameters:
      - description: Task payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/TaskMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/_ResponseWithData'
            - properties:
                data:
                  $ref: '#/definitions/hackathon-back_internal_model.Request'
              type: object
        "400":
          description: Invalid JSON body
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
        "500":
          description: Failed to create request
          schema:
            $ref: '#/definitions/_ResponseWithMessage'
      summary: Создать задачу сетевых проверок.
      tags:
      - Checks
  /check/ws/check/{request_id}:
    get:
      description: Открывает WS и присылает актуальные результаты по request_id до
        завершения всех проверок.
      parameters:
      - description: Request UUID
        in: path
        name: request_id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Стрим результатов сетевых проверок по WebSocket.
      tags:
      - Checks
...
swagger: "2.0"